fix: signed-overflow UB in sched-queue NVS save and quiet-hours display (RFC-0274)
- main.cpp sched_queue save: replace (long)sendAtMs - (long)nowMs with
  (int32_t)(sendAtMs - nowMsU) — signed reinterpretation of unsigned subtraction
  avoids UB at millis() wrap and correctly preserves sign. Saves slot timestamps
  correctly even if millis() wraps while a scheduled SMS is queued (RFC-0274).
- main.cpp sched_queue load: use nowMsU2 (uint32_t) directly for sendAtMs
  reconstruction to avoid implicit signed cast.
- telegram_poller.cpp quietHoursWarning(): same int32_t-cast fix (display only).

+++ src/telegram_poller.cpp
@@ -75,7 +75,8 @@ static String quietHoursWarning(uint32_t sendAtMs, uint32_t nowMs,
     if (qStart < 0 || qEnd < 0 || !wallTimeFn) return String();
     long wallNow = wallTimeFn();
     if (wallNow <= 1000000000L) return String(); // NTP not synced
-    long deltaMs = (long)sendAtMs - (long)nowMs;
+    // RFC-0274: signed reinterpret of unsigned subtraction avoids UB at millis() wrap.
+    long deltaMs = (long)(int32_t)(sendAtMs - nowMs);
     long sendAtUnix = wallNow + deltaMs / 1000L;
     if (!isInQuietHoursAt(sendAtUnix, qStart, qEnd)) return String();
     char buf[48];
